Stabilize JIT compilation and fix weak type scalars
- Add stabilize_weak_types() to eliminate 3-step recompilation
- Pre-warm both phase1/phase2 JIT functions before training loop
- Use explicit dtype=jnp.float32 for gamma, ema_usage, aspect_logits, tau, alpha_sum
- Use jnp.float32() wrapper for lambda_sharp and lambda_entropy_eff in train loop
- Add XLA compilation cache (.jax_cache/) with JAX_COMPILATION_CACHE_DIR
- Add diag_recompile.py diagnostic script
- Simplify alpha_sum computation in trainer.py and train_tpu_1b.py

+++ src/model/assembly.py
@@ -39,7 +39,7 @@ class WeightAssembler(nnx.Module):
         self.b_base = nnx.Param(jnp.zeros(d_B, dtype=jnp.bfloat16))
 
         # LoRA-style residual scale — init 1.0 so pool contribution is immediately active
-        self.gamma = nnx.Param(jnp.array(1.0))
+        self.gamma = nnx.Param(jnp.array(1.0, dtype=jnp.float32))
 
         # LayerNorm after assembly
         self.norm = nnx.LayerNorm(d_B, rngs=rngs)
